Chore/uv migration (#238)
* docs: add design spec for uv + pyproject.toml migration

Design for replacing pip/requirements.txt in backend/ with a uv-managed
project: pyproject.toml as the single manifest, uv.lock committed, uv run
as the command entrypoint.

Also records two pre-existing problems found while scoping:
- backend/setup.py is dead code (reads a non-existent README.rst)
- .gitlab-ci.yml deploys via pipenv against a Pipfile deleted in 8fac0de

* docs: add implementation plan for uv migration

* fix: import AcceptInvitationSerializer from its defining module

Commit 66513e7 ran ruff's unused-import autofix over auth_serializers.py,
which stripped AcceptInvitationSerializer from the `from .serializers`
line. Ruff was right that the name was unused within that module, but it
was a re-export consumed by auth_views.py -- so the whole Django app
stopped importing:

    ImportError: cannot import name 'AcceptInvitationSerializer'
    from 'api.v1.recruiter.auth_serializers'

Import it from serializers.py, where the class is actually defined,
rather than restoring the re-export -- ruff would strip that again.

Also drops the no-op expression statement left at auth_views.py:474 by
the same autofix, which removed the assignment target but kept the
subscript.

* build: add uv project manifest and lockfile

backend/pyproject.toml becomes the single dependency manifest: 27 runtime
pins under [project.dependencies] and 8 under a PEP 735 [dependency-groups]
dev group. No [build-system] is declared, so uv installs dependencies
without building the project itself; tool.uv.package = false states that
intent explicitly.

All 34 existing pins carry over byte-identical. ruff==0.16.2 is the only
addition -- it was already in active use (.ruff_cache is gitignored, and
66513e7 rewrote 20+ files with it) but declared nowhere and installed
nowhere, which is how an unpinned autofix broke the app.

Parity verified: all 27 declared runtime deps resolve to identical
versions under uv export --no-dev.

Also corrects the spec and plan, which said 30 runtime / 37 total pins.
The real counts are 27 and 34.

* build: remove requirements.txt, dev-requirements.txt and dead setup.py

pyproject.toml is now the sole dependency manifest. Parity with the two
requirements files was verified in the previous commit before deleting
them.

setup.py was already dead code: it opens a README.rst that has never
existed in this repo (so it raises FileNotFoundError if invoked),
declares install_requires=[], and nothing references it. Its metadata
(name, version, description, author, license) is preserved in
pyproject.toml.

* ci: deploy with uv sync --locked instead of dead pipenv invocation

The deploy job ran `pip install pipenv && pipenv install -d` against a
Pipfile deleted in 8fac0de, so it was already broken independently of
this migration.

Three deliberate changes beyond the tooling swap:

- Drops `source /home/peeljobs/env/bin/activate`; uv manages backend/.venv.
- --locked rather than --frozen: --frozen uses the lockfile without
  validating it against pyproject.toml, so a forgotten `uv lock` would
  silently ship stale dependencies. --locked fails the pipeline instead.
- --no-dev keeps dev tooling off production. The previous `pipenv
  install -d` explicitly installed dev dependencies there.

NOTE: supervisor config on the live box is not in this repo and likely
still points into /home/peeljobs/env. It must be repointed at
backend/.venv/bin/gunicorn before the first uv deploy.

* docs: replace pip workflow with uv across setup and run instructions

README.md and SETUP.md setup blocks collapse from venv-create +
activate + pip install down to `uv sync`. All manage.py / manage_server.py
/ celery invocations gain the `uv run` prefix.

Also repoints the documented systemd units from /var/www/peeljobs/venv/bin
to /var/www/peeljobs/backend/.venv/bin, and their WorkingDirectory to the
backend/ subdirectory where jobsp.wsgi actually resolves. Those units had
the same stale-path problem as the live-box supervisor config.

settings_local.py's "Install with: pip install -r dev-requirements.txt"
comments become "Installed by: uv sync" -- both apps are in the dev
group, which uv sync installs by default.

CLAUDE.md was updated the same way but is gitignored, so it is not part
of this commit.

.vscode/settings.json still names requirements.txt as a Copilot
instruction source. Left as-is per the spec's scope decision; it is
editor config, not project tooling.

* Implement code changes to enhance functionality and improve performance

* Refactor code for consistency and readability across multiple files

- Standardized string quotes to double quotes in various files.
- Removed unnecessary blank lines and spaces for cleaner code.
- Consolidated multi-line method calls into single lines where appropriate.
- Improved formatting in search and sitemap related files for better readability.
- Ensured consistent use of `prefetch_related` and `select_related` in querysets.
- Updated comments and docstrings for clarity.

* Refactor code to remove deprecated unicode_literals import, streamline form validation, and enhance readability

- Removed unnecessary unicode_literals import from multiple files.
- Updated super() calls to use the new style without specifying the class.
- Simplified form validation checks by replacing key existence checks with get() method.
- Improved import organization and removed unused imports across various modules.
- Enhanced readability of the code by restructuring conditional statements and loops.
- Ensured consistent use of list and dictionary methods for better performance and clarity.

* Refactor code for improved error handling, readability, and consistency across multiple files; add recruiter routes documentation.

* Add CSRF tokens to forms across various templates for enhanced security

- Updated view_microsite_page.html to include CSRF token in forms.
- Added CSRF token to edit_ticket.html for ticket editing.
- Included CSRF token in new_ticket.html for creating new tickets.
- Added CSRF token to view_ticket.html for comment submission.
- Included CSRF token in mobile_verify.html for mobile verification.
- Added CSRF token to payment_details.html for payment form.
- Included CSRF token in job listing and creation forms in list.html and new.html.
- Added CSRF token to advanced search filter templates.
- Included CSRF token in unsubscribe and unsubscribe_alerts.html for form submissions.

* Update Celery and Gunicorn configurations to use local settings module for improved environment management

+++ backend/dashboard/views/applicant_management.py
@@ -4,9 +4,9 @@ import re
 from datetime import datetime
 
 from django.db.models import Q
-from django.urls import reverse
 from django.http.response import HttpResponse, HttpResponseRedirect
 from django.shortcuts import get_object_or_404, render
+from django.urls import reverse
 
 from mpcomp.views import (
     get_prev_after_pages_count,
@@ -18,7 +18,6 @@ from peeldb.models import (
     Subscriber,
     User,
 )
-
 
 # Functions to move here from main views.py:
 
@@ -73,7 +72,7 @@ def applicants(request, status="all"):
         applicant = applicant.filter(date_joined__range=(start_date, end_date))
     applicant = applicant.order_by("-date_joined")
     items_per_page = 50
-    no_pages = int(math.ceil(float(applicant.count()) / items_per_page))
+    no_pages = math.ceil(float(applicant.count()) / items_per_page)
     if (
         "page" in request.POST
         and bool(re.search(r"[0-9]", request.POST.get("page")))
