+++ backend/dashboard/views/company_management.py
@@ -2,18 +2,18 @@ import json
 import math
 import re
 
-from django.urls import reverse
 from django.db.models import Q
 from django.http.response import HttpResponse, HttpResponseRedirect
 from django.shortcuts import get_object_or_404, render
 from django.template.defaultfilters import slugify
+from django.urls import reverse
 
+from mpcomp.aws import AWS
 from mpcomp.views import (
     get_aws_file_path,
     get_prev_after_pages_count,
     permission_required,
 )
-from mpcomp.aws import AWS
 from peeldb.models import (
     Company,
     JobPost,
@@ -26,8 +26,8 @@ from ..forms import (
     CompanyForm,
 )
 
-
 # Functions to move here from main views.py:
+
 
 @permission_required("activity_edit", "activity_view")
 def companies(request, company_type):
@@ -76,7 +76,7 @@ def companies(request, company_type):
         )
 
     items_per_page = 50
-    no_pages = int(math.ceil(float(companies.count()) / items_per_page))
+    no_pages = math.ceil(float(companies.count()) / items_per_page)
     companies = companies.order_by("-registered_date")
     page = request.POST.get("page") or request.GET.get("page")
     if page and bool(re.search(r"[0-9]", page)) and int(page) > 0:
@@ -110,7 +110,6 @@ def companies(request, company_type):
             "admin": request.GET.get("admin") if request.GET.get("admin") else "",
         },
     )
-
 
 
 @permission_required("activity_edit", "activity_view")
@@ -199,12 +198,10 @@ def edit_company(request, company_id):
     return render(request, "dashboard/company/new_company.html", {"company": company})
 
 
-
 @permission_required("activity_edit", "activity_view")
 def view_company(request, company_id):
     company = get_object_or_404(Company, id=company_id)
     return render(request, "dashboard/company/view.html", {"company": company})
-
 
 
 @permission_required("activity_edit")
@@ -300,8 +297,6 @@ def enable_paid_company(request, company_id):
     )
 
 
-
-
 @permission_required("activity_edit", "activity_view")
 def company_recruiters(request, company_id, status):
     company = get_object_or_404(Company, id=company_id)
@@ -311,7 +306,7 @@ def company_recruiters(request, company_id, status):
     else:
         recruiters = recruiters.filter(is_active=False)
     items_per_page = 100
-    no_pages = int(math.ceil(float(recruiters.count()) / items_per_page))
+    no_pages = math.ceil(float(recruiters.count()) / items_per_page)
     page = request.GET.get("page")
     if page and bool(re.search(r"[0-9]", page)) and int(page) > 0:
         if int(page) > (no_pages + 2):
@@ -346,7 +341,7 @@ def company_jobposts(request, company_id):
     company = get_object_or_404(Company, id=company_id)
     job_posts = company.get_jobposts()
     items_per_page = 100
-    no_pages = int(math.ceil(float(job_posts.count()) / items_per_page))
+    no_pages = math.ceil(float(job_posts.count()) / items_per_page)
     page = request.GET.get("page")
     if page and bool(re.search(r"[0-9]", page)) and int(page) > 0:
         if int(page) > (no_pages + 2):
@@ -381,7 +376,7 @@ def company_tickets(request, company_id):
     company = get_object_or_404(Company, id=company_id)
     items_per_page = 100
     tickets = company.get_company_tickets()
-    no_pages = int(math.ceil(float(tickets.count()) / items_per_page))
+    no_pages = math.ceil(float(tickets.count()) / items_per_page)
     page = request.GET.get("page")
     if page and bool(re.search(r"[0-9]", page)) and int(page) > 0:
         if int(page) > (no_pages + 2):
@@ -411,8 +406,6 @@ def company_tickets(request, company_id):
     )
 
 
-
-
 @permission_required("activity_edit")
 def edit_menu(request, menu_id, company_id):
     company = get_object_or_404(Company, id=company_id)
@@ -430,7 +423,6 @@ def edit_menu(request, menu_id, company_id):
         return HttpResponse(json.dumps(data))
 
 
-
 @permission_required("activity_edit")
 def delete_menu(request, menu_id, company_id):
     company = get_object_or_404(Company, id=company_id)
@@ -442,8 +434,6 @@ def delete_menu(request, menu_id, company_id):
     else:
         data = {"error": True, "response": "Some Problem Occurs"}
     return HttpResponse(json.dumps(data))
-
-
 
 
 @permission_required("activity_edit")
@@ -458,7 +448,6 @@ def menu_status(request, menu_id, company_id):
             menu.status = True
         menu.save()
     return HttpResponseRedirect(request.META.get("HTTP_REFERER"))
-
 
 
 @permission_required("activity_edit")
@@ -484,8 +473,6 @@ def menu_order(request, company_id):
     menu.lvl = current
     menu.save()
     return HttpResponseRedirect(request.META.get("HTTP_REFERER"))
-
-
 
 
 @permission_required("activity_edit")
